new features added

+++ client/src/components/Viewbook.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import { UseAuth } from './UseAuth';
+import { getGlobalEmail } from '../Global/global';
 
 import '../styles/viewbook.scss'; // Import your CSS file
 
@@ -13,7 +14,7 @@ function Viewbook() {
 
   useEffect(() => {
     // Fetch data from your API endpoint using Axios
-    axios.get("https://sachin2276.onrender.com/view") // Replace with your actual API endpoint
+    axios.get("http://localhost:3002/view") // Replace with your actual API endpoint
       .then((response) => {
         setBooks(response.data);
       })
@@ -28,10 +29,14 @@ function Viewbook() {
   };
 
   const handleDeleteClick = (book) => {
+    if(book.email != getGlobalEmail()) return;
     const confirmDelete = window.confirm(`Are you sure you want to delete "${book.title}"?`);
 
     if (confirmDelete) {
       // Send a DELETE request to your API to delete the book
+            
+  
+
       axios.delete(`http://localhost:3002/view/${book._id}`) // Replace with your delete endpoint
         .then(() => {
           // Remove the book from the state
@@ -50,6 +55,7 @@ function Viewbook() {
 
   const handleSaveEdit = (editedBook) => {
     // Send a PUT request to your API to update the book
+    if(editedBook.email != getGlobalEmail()) return;
     axios.put(`http://localhost:3002/view/${editedBook._id}`, editedBook) // Replace with your update endpoint
       .then(() => {
         // Update the book in the state
